LWE move to bytes db

+++ lwe_test.go
@@ -43,7 +43,7 @@ func retrieveBlocksLWE(t *testing.T, db *database.LWE, params *utils.ParamsLWE, 
 
 		res, err := c.ReconstructBytes(a)
 		require.NoError(t, err)
-		require.Equal(t, db.Matrix.Get(utils.VectorToMatrixIndices(i, db.Info.NumColumns)), res)
+		require.Equal(t, uint32(db.Matrix.Get(utils.VectorToMatrixIndices(i, db.Info.NumColumns))), res)
 	}
 	fmt.Printf("TotalCPU time %s: %.1fms\n", testName, totalTimer.Record())
